uncompressed: pulled AbstractDecoder into its own file
